Day 9: solved
The 4-neighborhood DFS (deep-first search) could have been used to search the
local minima of part 1.

Instead, it uses a homemade algorithm that roams the grid 4 times, one in each
direction:
- from left to right
- from right to left
- from top to bottom
- from bottom to top
Each one of these passage look for an increasing slope (the next value being
higher that the previous one), and marks it.
After for passages, we know the increasing slopes in all four direction, and are
able to know all the local minima of the grid

The DFS is probably more efficient