Day 12: re-write part 1
2 optimization:
- Use dictionary instead of class.
- Redo graph crossing. Instead of following one path after the other until
completion or failure, update each possible path one step further and filter out
the irrelevant paths.

Instead of using a recursion, do it with iteration. Each branching in the path
will only be a new line in the list of possible path, that will be processed on
next iteration, instead of creating a new recursion branch.