day8: rewrite with graph
Try to use a naive version of a (cyclic) graph, based on Rc + RefCell.
Previous version used two Vec to hold the reference to each node, and the links
between the nodes, and relied on a bit of back-and-forth between the two to walk
the graph.

Now it uses only one Vec of structs that have links to the other, so we can pass
from one node to the net directly.

TODO: because of the RefCell borrowing lifetime, I use clone() to get (and keep)
the next node, which may not be the more efficient.

Parsing a graph construction is longer, but walk is faster